requirements.txt
asgiref==3.8.1
celery==5.4.0
coverage==7.5.1
Django==4.2.2
django-celery-beat==2.6.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
redis==5.0.4
sqlparse==0.5.0